Unit tests fixed and xaml comments added
