+++ TestProject/TestsUpdater/TestClientViewModel.cs
@@ -1,16 +1,47 @@
-﻿using Moq;
+﻿/******************************************************************************
+* Filename    = TestClientViewModel.cs
+*
+* Author      = Garima Ranjan
+*
+* Product     = Updater
+* 
+* Project     = Lab Monitoring Software
+*
+* Description = Unit Tests for ClientViewModel.cs
+*****************************************************************************/
+
+using Moq;
 using Updater;
 using ViewModel.UpdaterViewModel;
 
+/// <summary>
+/// Unit Tests for the ClientViewModel class.
+/// Verifies the functionality and integration with the LogServiceViewModel and Client classes.
+/// </summary>
 namespace TestsUpdater;
 
 [TestClass]
 public class TestClientViewModel
 {
+    /// <summary>
+    /// Mocked instance of the LogServiceViewModel used to verify interactions with the logging system.
+    /// </summary>
     private Mock<LogServiceViewModel>? _mockLogServiceViewModel;
+
+    /// <summary>
+    /// Instance of the ClientViewModel being tested.
+    /// </summary>
     private ClientViewModel? _viewModel;
+
+    /// <summary>
+    /// Mocked instance of the Client class, created using reflection.
+    /// </summary>
     private Client? _mockClient;
 
+    /// <summary>
+    /// Initializes the test environment before each test method runs.
+    /// Sets up mock objects and injects dependencies into the ClientViewModel.
+    /// </summary>
     [TestInitialize]
     public void Setup()
     {
@@ -20,7 +51,9 @@ public class TestClientViewModel
         // Use reflection to instantiate the Client class
         System.Reflection.ConstructorInfo? constructorInfo = typeof(Client).GetConstructor(
             System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.NonPublic,
-            null, [], null);
+            null,
+            Array.Empty<Type>(),
+            null);
 
         if (constructorInfo == null)
         {
@@ -34,7 +67,7 @@ public class TestClientViewModel
             .GetField("OnLogUpdate", System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic)
             ?.SetValue(null, (Action<string>)_mockLogServiceViewModel.Object.UpdateLogDetails);
 
-        // Initializing ClientViewModel and inject the mock LogServiceViewModel
+        // Initializing ClientViewModel and injecting the mock LogServiceViewModel
         _viewModel = new ClientViewModel(_mockLogServiceViewModel.Object);
 
         // Injecting the private s_client field in ClientViewModel with our mock
@@ -43,21 +76,24 @@ public class TestClientViewModel
             ?.SetValue(_viewModel, _mockClient);
     }
 
+    /// <summary>
+    /// Tests that the SyncUpAsync method invokes the client's sync-up logic
+    /// and logs the completion of the operation.
+    /// </summary>
     [TestMethod]
     public async Task TestSyncUpAsyncInvokesClientAndLogsCompletion()
     {
         bool syncUpCalled = false;
 
-        // Using reflection to mock SyncUp method behavior
+        // Mock the SyncUp method behavior
         typeof(Client)
             .GetMethod("SyncUp", System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.Public)
-            ?.Invoke(_mockClient, []);
+            ?.Invoke(_mockClient, Array.Empty<object>());
 
         if (_mockLogServiceViewModel == null)
         {
             Assert.Fail("_mockLogServiceViewModel is not initialized.");
         }
-
 
         // Update LogServiceViewModel mock
         _mockLogServiceViewModel
@@ -72,7 +108,7 @@ public class TestClientViewModel
 
         await _viewModel.SyncUpAsync();
 
-
+        // Assert that the sync-up was called and the log was updated
         Assert.IsTrue(syncUpCalled, "SyncUp should have been called.");
         _mockLogServiceViewModel.Verify(
             log => log.UpdateLogDetails("Sync completed."),
@@ -81,6 +117,9 @@ public class TestClientViewModel
         );
     }
 
+    /// <summary>
+    /// Verifies that the PropertyChanged event is raised when a property changes.
+    /// </summary>
     [TestMethod]
     public void TestOnPropertyChangedEventIsRaised()
     {
@@ -93,10 +132,12 @@ public class TestClientViewModel
             raisedPropertyName = args.PropertyName;
         };
 
+        // Trigger the PropertyChanged event
         _viewModel.GetType()
             .GetMethod("OnPropertyChanged", System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.NonPublic)
-            ?.Invoke(_viewModel, ["TestProperty"]);
+            ?.Invoke(_viewModel, new object[] { "TestProperty" });
 
+        // Assert that the event was raised with the correct property name
         Assert.IsTrue(eventRaised, "PropertyChanged event should be raised.");
         Assert.AreEqual("TestProperty", raisedPropertyName, "PropertyChanged event should be raised with the correct property name.");
     }
