+++ TestProject/TestsUpdater/TestDirectoryMetadataComparer.cs
@@ -1,7 +1,15 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
+﻿/******************************************************************************
+* Filename    = TestDirectoryMetadataComparer.cs
+*
+* Author      = Garima Ranjan & Amithabh A.
+*
+* Product     = Updater
+* 
+* Project     = Lab Monitoring Software
+*
+* Description = Unit Tests for DirectoryMetadataComparer.cs
+*****************************************************************************/
+
 using Updater;
 
 namespace TestsUpdater;
@@ -9,117 +17,142 @@ namespace TestsUpdater;
 [TestClass]
 public class TestDirectoryMetadataComparer
 {
+    /// <summary>
+    /// Sample metadata set A for testing.
+    /// </summary>
     private List<FileMetadata>? _metadataA;
+
+    /// <summary>
+    /// Sample metadata set B for testing.
+    /// </summary>
     private List<FileMetadata>? _metadataB;
 
+    /// <summary>
+    /// Sets up sample file metadata for testing.
+    /// This method runs before each test case.
+    /// </summary>
     [TestInitialize]
     public void Setup()
     {
-        // Prepare some sample metadata for testing
-        _metadataA =
-        [
-            new () { FileName = "file1.txt", FileHash = "hash1" },
-            new () { FileName = "file2.txt", FileHash = "hash2" },
-            new () { FileName = "file5.txt", FileHash = "hash5" },
-        ];
+        // Prepare sample metadata for two directories
+        _metadataA = new List<FileMetadata>
+        {
+            new() { FileName = "file1.txt", FileHash = "hash1" },
+            new() { FileName = "file2.txt", FileHash = "hash2" },
+            new() { FileName = "file5.txt", FileHash = "hash5" },
+        };
 
-        _metadataB =
-        [
-            new () { FileName = "file2.txt", FileHash = "hash2" },
-            new () { FileName = "file3.txt", FileHash = "hash3" },
-            new (){ FileName = "file4.txt", FileHash = "hash1" },
-        ];
+        _metadataB = new List<FileMetadata>
+        {
+            new() { FileName = "file2.txt", FileHash = "hash2" },
+            new() { FileName = "file3.txt", FileHash = "hash3" },
+            new() { FileName = "file4.txt", FileHash = "hash1" },
+        };
     }
 
+    /// <summary>
+    /// Verifies that CompareMetadata identifies differences correctly between metadata sets A and B.
+    /// </summary>
     [TestMethod]
     public void TestCompareMetadataShouldIdentifyDifferences()
     {
-        // Arrange
+        // Arrange: Create an instance of the comparer
         var comparer = new DirectoryMetadataComparer(_metadataA!, _metadataB!);
 
-        // Act
+        // Act: Retrieve differences between the two metadata sets
         List<MetadataDifference> differences = comparer.Differences;
 
-        // Assert: Check if differences are as expected
-        Assert.AreEqual(1, differences.First(d => d.Key == "-1").Value.Count);
-        Assert.AreEqual(1, differences.First(d => d.Key == "0").Value.Count);
-        Assert.AreEqual(1, differences.First(d => d.Key == "1").Value.Count);
+        // Assert: Verify differences match expectations
+        Assert.AreEqual(1, differences.First(d => d.Key == "-1").Value.Count, "Expected 1 file missing in A.");
+        Assert.AreEqual(1, differences.First(d => d.Key == "0").Value.Count, "Expected 1 file common between A and B.");
+        Assert.AreEqual(1, differences.First(d => d.Key == "1").Value.Count, "Expected 1 file missing in B.");
     }
 
+    /// <summary>
+    /// Verifies that CheckForRenamesAndMissingFiles identifies files missing in metadata set A.
+    /// </summary>
     [TestMethod]
     public void TestCheckForRenamesAndMissingFilesShouldIdentifyMissingFilesInA()
     {
-        // Arrange
+        // Arrange: Create an instance of the comparer
         var comparer = new DirectoryMetadataComparer(_metadataA!, _metadataB!);
 
-        // Act
+        // Act: Retrieve unique client-side files
         List<string> uniqueClientFiles = comparer.UniqueClientFiles;
 
-        // Assert: Check if missing files from A are identified
-        Assert.AreEqual(1, uniqueClientFiles.Count); // file4.txt is only in B
-        Assert.AreEqual("file3.txt", uniqueClientFiles.First());
+        // Assert: Verify missing files in A are identified correctly
+        Assert.AreEqual(1, uniqueClientFiles.Count, "Expected 1 file missing in A.");
+        Assert.AreEqual("file3.txt", uniqueClientFiles.First(), "Expected file3.txt to be identified as missing.");
     }
 
+    /// <summary>
+    /// Verifies that CheckForOnlyInAFiles identifies files missing in metadata set B.
+    /// </summary>
     [TestMethod]
     public void TestCheckForOnlyInAFilesShouldIdentifyMissingFilesInB()
     {
-        // Arrange
+        // Arrange: Create an instance of the comparer
         var comparer = new DirectoryMetadataComparer(_metadataA!, _metadataB!);
 
-        // Act
+        // Act: Retrieve unique server-side files
         List<string> uniqueServerFiles = comparer.UniqueServerFiles;
 
-        // Assert: Check if missing files from B are identified
-        Assert.AreEqual(1, uniqueServerFiles.Count); // file1.txt is only in A
-        Assert.AreEqual("file5.txt", uniqueServerFiles.First());
+        // Assert: Verify missing files in B are identified correctly
+        Assert.AreEqual(1, uniqueServerFiles.Count, "Expected 1 file missing in B.");
+        Assert.AreEqual("file5.txt", uniqueServerFiles.First(), "Expected file5.txt to be identified as missing.");
     }
 
+    /// <summary>
+    /// Verifies that CheckForSameNameDifferentHash identifies files with hash mismatches.
+    /// </summary>
     [TestMethod]
     public void TestCheckForSameNameDifferentHashShouldIdentifyFileHashMismatch()
     {
-        // Arrange
-        // Using null-forgiving operator if you are confident they are initialized in Setup
+        // Arrange: Create an instance of the comparer
         var comparer = new DirectoryMetadataComparer(_metadataA!, _metadataB!);
 
-        // Act
-        // Ensure invalidSyncUpFiles is non-null
-        List<string> invalidSyncUpFiles = comparer.InvalidSyncUpFiles ?? [];
+        // Act: Retrieve files with hash mismatches
+        List<string> invalidSyncUpFiles = comparer.InvalidSyncUpFiles ?? new List<string>();
 
-        // Assert
-        Assert.AreEqual(0, invalidSyncUpFiles.Count);
+        // Assert: Verify that no files have hash mismatches
+        Assert.AreEqual(0, invalidSyncUpFiles.Count, "Expected no hash mismatches.");
     }
 
-
-
+    /// <summary>
+    /// Verifies that ValidateSync returns true when no invalid files exist.
+    /// </summary>
     [TestMethod]
     public void TestValidateSyncShouldReturnTrueWhenNoInvalidFilesExist()
     {
-        // Arrange
+        // Arrange: Create an instance of the comparer
         var comparer = new DirectoryMetadataComparer(_metadataA!, _metadataB!);
 
-        // Act
+        // Act: Check if synchronization is valid
         bool canSync = comparer.ValidateSync();
 
-        // Assert: Should return true due to file3.txt having different hashes
-        Assert.IsTrue(canSync);
+        // Assert: Verify synchronization is allowed
+        Assert.IsTrue(canSync, "Expected synchronization to be valid.");
     }
 
+    /// <summary>
+    /// Verifies that ValidateSync returns false when invalid files exist.
+    /// </summary>
     [TestMethod]
-    public void TestValidateSyncShouldReturnFalseWhenNoInvalidFiles()
+    public void TestValidateSyncShouldReturnFalseWhenInvalidFilesExist()
     {
-        // Arrange
+        // Arrange: Modify metadataB to introduce invalid files
         var metadataBInvalidFiles = new List<FileMetadata>
         {
-            new () { FileName = "file1.txt", FileHash = "hash1" },
-            new () { FileName = "file2.txt", FileHash = "hash2" },
-            new () { FileName = "file5.txt", FileHash = "hash7" }
+            new() { FileName = "file1.txt", FileHash = "hash1" },
+            new() { FileName = "file2.txt", FileHash = "hash2" },
+            new() { FileName = "file5.txt", FileHash = "hash7" }
         };
         var comparer = new DirectoryMetadataComparer(_metadataA!, metadataBInvalidFiles);
 
-        // Act
+        // Act: Check if synchronization is valid
         bool canSync = comparer.ValidateSync();
 
-        // Assert: Should return false as there are no invalid files
-        Assert.IsFalse(canSync);
+        // Assert: Verify synchronization is disallowed
+        Assert.IsFalse(canSync, "Expected synchronization to be invalid due to hash mismatches.");
     }
 }
